# eric-am-common-wfs-server-api/src/main/resources/mainYaml/api.yaml
...
openapi: 3.0.3
info:
  version: 0.0.1
  title: Common Workflow Service
  description: >
    Common Workflow Service
servers:
  - url: '{protocol}://{hostname}/{basePath}'
    variables:
      protocol:
        default: https
      hostname:
        default: am-common-wfs
      basePath:
        default: api
...
paths:
...
  /lcm/v3/resources/{releaseName}:
    get:
      tags:
        - Lcm
      operationId: getHistoryV3Resource
      parameters:
        - name: releaseName
          in: path
          description: |
            Identifier of the resource.
          schema:
            type: string
            pattern: '[a-z]([-a-z0-9]*[a-z0-9])?'
          x-pattern-message: 'RELEASE_NAME_ERROR_MSG'
          required: true
        - name: instanceId
          description: >
            id of the instance to query
          in: query
          schema:
            type: string
      responses:
        '200':
          description: >
            OK - Information about the history of the resource
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
...
  /internal/v2/namespaces/{namespace}/delete:
    post:
      tags:
        - Internal
      operationId: deleteNamespace
      parameters:
        - name: Idempotency-key
          description: |
            Unique request identifier which is generated automatically.
          in: header
          required: true
          schema:
            type: string
        - name: namespace
          in: path
          description: |
            namespace of the resource.
          schema:
            type: string
          required: true
        - name: releaseName
          in: query
          description: |
            Identifier of the resource.
          schema:
            type: string
            pattern: '[a-z]([-a-z0-9]*[a-z0-9])?'
          x-pattern-message: 'RELEASE_NAME_ERROR_MSG'
          required: true
        - name: applicationTimeOut
          in: query
          description: |
            Maximum time allocated for application instantiation.
          schema:
            type: string
        - name: clusterName
          in: query
          description: |
            Cluster name where the resource is located.
          schema:
            type: string
            pattern: '[0-9a-zA-Z][0-9a-zA-Z-_]*(\.config)?$'
          x-pattern-message: 'CLUSTER_NAME_ERROR_MSG'
          required: true
        - name: lifecycleOperationId
          in: query
          description: |
            the lifecycle operation id related to this request.
          schema:
            type: string
            minLength: 1
          required: true
          x-message: "LIFECYCLE_OPERATION_ID_BLANK_EMPTY"
          x-required-message: 'LIFECYCLE_OPERATION_ID_NULL'
      responses:
        '202':
          $ref: '#/components/responses/AcceptedNoBody'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /internal/kubernetes/pods/scale/down:
    post:
      tags:
        - Internal
      operationId: scaleDown
      parameters:
        - name: Idempotency-key
          description: |
            Unique request identifier which is generated automatically.
          in: header
          required: true
          schema:
            type: string
      requestBody:
        description: The scale down resource parameters
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalScaleInfo'
      responses:
        '202':
          $ref: '#/components/responses/AcceptedNoBody'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /internal/cluster/validate:
    post:
      tags:
        - Internal
      operationId: clusterConfigFileValidate
      description: |
        Internal api which is called from LCM to validate cluster config file.
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - clusterConfig
              properties:
                clusterConfig:
                  description: Cluster config file.
                  type: string
                  format: binary
      responses:
        '200':
          description: >
            OK
            Cluster config is validated successfully. The response body shall contain cluster host url.
          headers:
            Content-Type:
              description: The MIME type of the body of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterServerDetailsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /internal/kubernetes/pvcs/{releaseName}/delete:
    post:
      tags:
        - Internal
      operationId: deletePvcs
      parameters:
        - name: Idempotency-key
          description: |
            Unique request identifier which is generated automatically.
          in: header
          required: true
          schema:
            type: string
        - name: releaseName
          in: path
          description: |
            Identifier of the resource.
          schema:
            type: string
            pattern: '[a-z]([-a-z0-9]*[a-z0-9])?'
          x-pattern-message: 'RELEASE_NAME_ERROR_MSG'
          required: true
        - name: applicationTimeOut
          in: query
          description: |
            Maximum time allocated for application instantiation.
          schema:
            type: string
        - name: clusterName
          in: query
          description: |
            Cluster name where the resource is located.
          schema:
            type: string
            pattern: '[0-9a-zA-Z][0-9a-zA-Z-_]*(\.config)?$'
          x-pattern-message: 'CLUSTER_NAME_ERROR_MSG'
        - name: lifecycleOperationId
          in: query
          description: |
            the lifecycle operation id related to this request.
          schema:
            type: string
            minLength: 1
          required: true
          x-message: "LIFECYCLE_OPERATION_ID_BLANK_EMPTY"
          x-required-message: 'LIFECYCLE_OPERATION_ID_NULL'
        - name: state
          in: query
          description: |
            the current lifecycle state of the lifecycle operation id.
          schema:
            type: string
            minLength: 1
          required: true
          x-message: "STATE_BLANK_EMPTY"
          x-required-message: 'STATE_NULL'
        - name: namespace
          in: query
          description: |
            The namespace the resource is located in.
          schema:
            type: string
            minLength: 1
          required: true
          x-message: "NAMESPACE_EMPTY"
          x-required-message: 'NAMESPACE_NULL'
        - name: labels
          in: query
          description: |
            List of labels for which pvcs need to be deleted
          schema:
            type: array
            items:
              type: string
      responses:
        '202':
          $ref: '#/components/responses/AcceptedNoBody'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /internal/kubernetes/secrets:
    parameters:
      - name: namespace
        in: query
        description: |
          The namespace the resource is located in.
        schema:
          type: string
          minLength: 1
        required: true
        x-message: "NAMESPACE_EMPTY"
        x-required-message: 'NAMESPACE_NULL'
      - name: clusterName
        in: query
        description: |
          Cluster name where the resource is located.
        schema:
          type: string
          pattern: '[0-9a-zA-Z][0-9a-zA-Z-_]*(\.config)?$'
        x-pattern-message: 'CLUSTER_NAME_ERROR_MSG'
      - name: fetchTimeOut
        in: query
        description: |
          The fetchTimeOut is the time needed to fetch the secret details from kubernetes
        schema:
          type: string
        required: false
    post:
      tags:
        - Internal
      operationId: getSecretsInNamespaceWithMultipart
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                clusterConfig:
                  type: string
                  format: binary
      responses:
        '200':
          description: >
            OK
            Information about zero or more VNF instances was queried
            successfully. The response body shall contain representations of
            zero or more VNF instances.
          headers:
            Content-Type:
              description: The MIME type of the body of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                description: Map of secret name and secret data
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Internal
      operationId: getSecretsInNamespace
      responses:
        '200':
          description: >
            Accepted
            The request was accepted for processing, but the  processing has not
            been completed. The response body shall be empty. The HTTP response
            shall include a "Location" HTTP header that contains the URI of the
            newly-created "VNF LCM operation occurrence" resource corresponding
            to the operation.
          headers:
            Content-Type:
              description: The MIME type of the body of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                description: Map of secret name and secret data
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /internal/kubernetes/secrets/{secretName}:
    put:
      tags:
        - Internal
      operationId: patchSecretInNamespace
      parameters:
        - name: secretName
          in: path
          description: |
            Name of the secret.
          schema:
            type: string
          required: true
      requestBody:
        description: Parameters to patch a Secret
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretInfo'
      responses:
        '202':
          $ref: '#/components/responses/AcceptedNoBody'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /internal/kubernetes/values/{releaseName}:
    parameters:
      - name: releaseName
        in: path
        description: |
          Identifier of the resource.
        required: true
        schema:
          type: string
          pattern: '[a-z]([-a-z0-9]*[a-z0-9])?'
        x-pattern-message: 'RELEASE_NAME_ERROR_MSG'
      - name: namespace
        in: query
        description: |
          The namespace the resource is located in.
        schema:
          type: string
          minLength: 1
        required: true
        x-message: "NAMESPACE_EMPTY"
        x-required-message: 'NAMESPACE_NULL'
      - name: clusterName
        in: query
        description: |
          The cluster the resource is located in.
        schema:
          type: string
          minLength: 1
          pattern: '[0-9a-zA-Z][0-9a-zA-Z-_]*(\.config)?$'
        required: true
        x-pattern-message: 'CLUSTER_NAME_ERROR_MSG'
      - name: fetchTimeOut
        in: query
        description: |
          The fetchTimeOut is the time needed to fetch the values from kubernetes
        schema:
          type: string
        required: false
    post:
      tags:
        - Internal
      operationId: getValuesInReleaseWithMultipart
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                clusterConfig:
                  type: string
                  format: binary
      responses:
        '200':
          description: >
            OK
            Information about values file was queried
            successfully.
          headers:
            Content-Type:
              description: The MIME type of the body of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                description: Map of helm values
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Internal
      operationId: getValuesInRelease
      responses:
        '200':
          description: >
            OK
            Information about values file was queried
            successfully.
          headers:
            Content-Type:
              description: The MIME type of the body of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                description: Map of helm values
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /internal/pods:
    post:
      tags:
        - Internal Resource Controller
      description: Retrieves status of the pods by release names and cluster name.
      operationId: getPodStatusWithMultipart
      parameters:
        - name: clusterName
          in: query
          description: Cluster name where the resource is located.
          schema:
            type: string
            pattern: '[0-9a-zA-Z][0-9a-zA-Z-_]*(\.config)?$'
          x-pattern-message: 'CLUSTER_NAME_ERROR_MSG'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                clusterConfig:
                  type: string
                  format: binary
                json:
                  type: string
      responses:
        '200':
          description: >
            OK
            Cluster config is validated successfully. The response body shall contain cluster host url.
          headers:
            Content-Type:
              description: The MIME type of the body of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: 'PodStatusResponseList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
...
    AcceptedNoBody:
      description: >
        Accepted
        The request was accepted for processing, but the processing has not
        been completed. The response body shall be empty.
    BadRequest:
      description: >
        Bad Request
        If the request is malformed or syntactically incorrect (e.g. if the request URI
        contains incorrect query parameters or the payload body contains a syntactically
        incorrect data structure), the API producer shall respond with this response code.
        ---
        If the response to a GET request which queries a container resource would be so big
        that the performance of the API producer is adversely affected, and the API
        producer does not support paging for the affected resource, it shall respond with this
        response code.
        ---
        If there is an application error related to the client's input that
        cannot be easily mapped to any other HTTP response code ("catch all
        error"), the API producer shall respond with this response code.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetails'
    NotFound:
      description: >
        Not Found
        If the API producer did not find a current representation for the
        resource addressed by the URI passed in the request, or is not
        willing to disclose that one exists, it shall respond with this
        response code.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetails'
    InternalServerError:
      description: >
        Internal Server Error
        If there is an application error not related to the client's input
        that cannot be easily mapped to any other HTTP response code ("catch
        all error"), the API producer shall respond with this response code.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetails'
  schemas:
    ErrorDetails:
      type: object
      properties:
        errorDetails:
          type: object
...